fix: use bigint for sort_order to support Date.now() values

+++ src/hooks.server.ts
@@ -10,7 +10,7 @@ async function runMigrations() {
 			user_id text NOT NULL REFERENCES users(id) ON DELETE CASCADE,
 			category_id text NOT NULL REFERENCES vault_categories(id) ON DELETE CASCADE,
 			name text NOT NULL,
-			sort_order integer NOT NULL DEFAULT 0,
+			sort_order bigint NOT NULL DEFAULT 0,
 			created_at timestamp DEFAULT now() NOT NULL
 		)
 	`;
